feat(064): split game screen by concerns
