+++ app/src/main/java/org/cescfe/numpairs/ui/screen/GameScreen.kt
@@ -1,118 +1,27 @@
 package org.cescfe.numpairs.ui.screen
 
-import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.BorderStroke
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.heightIn
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.widthIn
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.selection.selectableGroup
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.AlertDialog
-import androidx.compose.material3.Button
-import androidx.compose.material3.ColorScheme
-import androidx.compose.material3.DropdownMenu
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.lerp
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.selected
-import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.semantics.stateDescription
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.zIndex
 import org.cescfe.numpairs.R
 import org.cescfe.numpairs.domain.puzzle.Operator
-import org.cescfe.numpairs.domain.puzzle.PuzzleCompletionState
 import org.cescfe.numpairs.initialPuzzle
-import org.cescfe.numpairs.ui.components.AvailableNumberChip
-import org.cescfe.numpairs.ui.components.AvailableNumberChipStyle
-import org.cescfe.numpairs.ui.components.PuzzleTile
 import org.cescfe.numpairs.ui.theme.NumPairsTheme
 
-private const val BOARD_MAX_VISUAL_COLUMN_COUNT = 4
-private val BOARD_TILE_MIN_WIDTH = 112.dp
-private val BOARD_TILE_MAX_WIDTH = 144.dp
-private val BOARD_TILE_SPACING = 12.dp
-private val STRIP_CHIP_SPACING = 4.dp
-private val STRIP_HORIZONTAL_PADDING = 8.dp
-private val STRIP_VERTICAL_PADDING = 14.dp
-private val TILE_OPERATOR_MENU_CORNER_RADIUS = 16.dp
-private val TILE_OPERATOR_MENU_PADDING = 8.dp
-private val TILE_OPERATOR_MENU_OPTION_SPACING = 8.dp
-private val TILE_OPERATOR_MENU_OPTION_HORIZONTAL_PADDING = 12.dp
-private val TILE_OPERATOR_MENU_OPTION_VERTICAL_PADDING = 8.dp
-private val TILE_OPERAND_SHEET_MAX_HEIGHT = 320.dp
-private val TILE_OPERAND_SHEET_PADDING = 20.dp
-private val TILE_OPERAND_SHEET_GRID_SPACING = 12.dp
-private val TILE_OPERAND_SHEET_OPTION_MIN_WIDTH = 72.dp
-private val TILE_OPERAND_SHEET_OPTION_CARD_MIN_WIDTH = 56.dp
-private val TILE_OPERAND_SHEET_OPTION_CARD_MAX_WIDTH = 76.dp
-private val TILE_OPERAND_SHEET_OPTION_MIN_HEIGHT = 50.dp
-private val TILE_OPERAND_SHEET_OPTION_CORNER_RADIUS = 18.dp
-private val TILE_OPERAND_HINT_OVERLAY_LIFT = 8.dp
-private val TILE_OPERAND_HINT_EDGE_INSET = 8.dp
-private val TILE_OPERAND_HINT_CORNER_RADIUS = 999.dp
-private val TILE_OPERAND_HINT_HORIZONTAL_PADDING = 6.dp
-private val TILE_OPERAND_HINT_VERTICAL_PADDING = 2.dp
-private val TILE_OPERAND_HINT_TEXT_WEIGHT = FontWeight.Medium
-private val PUZZLE_OUTCOME_CORNER_RADIUS = 24.dp
-private val PUZZLE_OUTCOME_HORIZONTAL_PADDING = 16.dp
-private val PUZZLE_OUTCOME_VERTICAL_PADDING = 14.dp
-private const val SUCCESS_OVERLAY_SCRIM_ALPHA = 0.44f
-private val SUCCESS_OVERLAY_CARD_MAX_WIDTH = 280.dp
-private val SUCCESS_OVERLAY_HORIZONTAL_PADDING = 24.dp
-private val SUCCESS_OVERLAY_VERTICAL_PADDING = 22.dp
-private val SUCCESS_OVERLAY_CARD_CORNER_RADIUS = 28.dp
-private val SUCCESS_OVERLAY_BADGE_SIZE = 44.dp
-private val SUCCESS_OVERLAY_SUCCESS_GREEN = Color(0xFF2E7D32)
-private val SUCCESS_OVERLAY_SUCCESS_GREEN_SOFT = Color(0xFFE6F4EA)
-
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun GameScreen(
     uiState: GameUiState,
@@ -197,719 +106,6 @@ fun GameScreen(
             onConfirm = onTileOperandSelectionConfirmed
         )
     }
-}
-
-@Composable
-private fun SuccessOverlay(onDismiss: () -> Unit) {
-    val interactionSource = remember { MutableInteractionSource() }
-
-    BackHandler(onBack = onDismiss)
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(MaterialTheme.colorScheme.scrim.copy(alpha = SUCCESS_OVERLAY_SCRIM_ALPHA))
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                onClick = onDismiss
-            )
-            .testTag(GameScreenTestTags.SUCCESS_OVERLAY),
-        contentAlignment = Alignment.Center
-    ) {
-        Surface(
-            modifier = Modifier
-                .padding(horizontal = 32.dp)
-                .widthIn(max = SUCCESS_OVERLAY_CARD_MAX_WIDTH),
-            shape = RoundedCornerShape(SUCCESS_OVERLAY_CARD_CORNER_RADIUS),
-            color = SUCCESS_OVERLAY_SUCCESS_GREEN_SOFT,
-            contentColor = SUCCESS_OVERLAY_SUCCESS_GREEN,
-            border = BorderStroke(width = 1.dp, color = SUCCESS_OVERLAY_SUCCESS_GREEN)
-        ) {
-            Column(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(
-                        horizontal = SUCCESS_OVERLAY_HORIZONTAL_PADDING,
-                        vertical = SUCCESS_OVERLAY_VERTICAL_PADDING
-                    ),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                Surface(
-                    shape = CircleShape,
-                    color = SUCCESS_OVERLAY_SUCCESS_GREEN.copy(alpha = 0.12f),
-                    contentColor = SUCCESS_OVERLAY_SUCCESS_GREEN
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .width(SUCCESS_OVERLAY_BADGE_SIZE)
-                            .heightIn(min = SUCCESS_OVERLAY_BADGE_SIZE),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        Text(
-                            text = "OK",
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                }
-                Text(
-                    text = stringResource(R.string.success_overlay_message),
-                    modifier = Modifier.testTag(GameScreenTestTags.SUCCESS_OVERLAY_MESSAGE),
-                    style = MaterialTheme.typography.headlineSmall,
-                    fontWeight = FontWeight.SemiBold
-                )
-                Text(
-                    text = stringResource(R.string.success_overlay_supporting_text),
-                    style = MaterialTheme.typography.bodyMedium
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun PuzzleOutcomeBanner(puzzleOutcome: PuzzleOutcomeUiState.Invalid, modifier: Modifier = Modifier) {
-    val colorScheme = MaterialTheme.colorScheme
-
-    Surface(
-        modifier = modifier.testTag(GameScreenTestTags.PUZZLE_OUTCOME),
-        shape = RoundedCornerShape(PUZZLE_OUTCOME_CORNER_RADIUS),
-        color = colorScheme.errorContainer,
-        contentColor = colorScheme.onErrorContainer,
-        border = BorderStroke(width = 1.dp, color = colorScheme.error)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    horizontal = PUZZLE_OUTCOME_HORIZONTAL_PADDING,
-                    vertical = PUZZLE_OUTCOME_VERTICAL_PADDING
-                ),
-            verticalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            Text(
-                text = stringResource(R.string.puzzle_outcome_invalid_title),
-                modifier = Modifier.testTag(GameScreenTestTags.PUZZLE_OUTCOME_TITLE),
-                style = MaterialTheme.typography.titleMedium,
-                fontWeight = FontWeight.SemiBold
-            )
-            Text(
-                text = puzzleOutcome.message(),
-                modifier = Modifier.testTag(GameScreenTestTags.PUZZLE_OUTCOME_MESSAGE),
-                style = MaterialTheme.typography.bodyMedium
-            )
-        }
-    }
-}
-
-@Composable
-private fun BoardSection(
-    tiles: List<TileUiState>,
-    onTileLeftOperandTapped: (Int) -> Unit,
-    onTileRightOperandTapped: (Int) -> Unit,
-    onTileOperatorTapped: (Int) -> Unit,
-    onTileResetTapped: (Int) -> Unit,
-    tileOperatorSelectionDialog: TileOperatorSelectionDialogUiState?,
-    onTileOperatorSelectionDismissed: () -> Unit,
-    onTileOperatorSelectionConfirmed: (Operator) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val boardContentDescription = stringResource(R.string.board_content_description)
-
-    BoxWithConstraints(
-        modifier = modifier
-            .testTag(GameScreenTestTags.BOARD)
-            .semantics {
-                contentDescription = boardContentDescription
-            }
-    ) {
-        val visualColumnCount = calculateBoardColumnCount(maxWidth)
-        val tileWidth = calculateBoardTileWidth(
-            availableWidth = maxWidth,
-            visualColumnCount = visualColumnCount
-        )
-        val visualRows = tiles.withIndex().toList().chunked(visualColumnCount)
-
-        Column(
-            verticalArrangement = Arrangement.spacedBy(BOARD_TILE_SPACING)
-        ) {
-            visualRows.forEach { row ->
-                Box(
-                    modifier = Modifier.fillMaxWidth(),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Row(
-                        horizontalArrangement = Arrangement.spacedBy(BOARD_TILE_SPACING)
-                    ) {
-                        row.forEach { indexedTile ->
-                            val tileIndex = indexedTile.index
-                            val tile = indexedTile.value
-                            val tileOperatorSelectionUiState = tileOperatorSelectionDialog
-                                ?.takeIf { dialogUiState -> dialogUiState.tileIndex == tileIndex }
-
-                            PuzzleTile(
-                                tile = tile,
-                                modifier = Modifier
-                                    .testTag(GameScreenTestTags.tile(tileIndex))
-                                    .width(tileWidth)
-                                    .wrapContentHeight(),
-                                leftOperandModifier = Modifier.testTag(GameScreenTestTags.tileLeftOperand(tileIndex)),
-                                leftOperandContentDescription = tileLeftOperandContentDescription(tile),
-                                onLeftOperandClick = { onTileLeftOperandTapped(tileIndex) },
-                                operatorModifier = Modifier.testTag(GameScreenTestTags.tileOperator(tileIndex)),
-                                operatorContentDescription = tileOperatorContentDescription(tile),
-                                onOperatorClick = { onTileOperatorTapped(tileIndex) },
-                                operatorOverlay = {
-                                    tileOperatorSelectionUiState?.let { dialogUiState ->
-                                        TileOperatorSelectionMenu(
-                                            dialogUiState = dialogUiState,
-                                            onDismiss = onTileOperatorSelectionDismissed,
-                                            onConfirm = onTileOperatorSelectionConfirmed
-                                        )
-                                    }
-                                },
-                                rightOperandModifier = Modifier.testTag(GameScreenTestTags.tileRightOperand(tileIndex)),
-                                rightOperandContentDescription = tileRightOperandContentDescription(tile),
-                                onRightOperandClick = { onTileRightOperandTapped(tileIndex) },
-                                resetModifier = Modifier.testTag(GameScreenTestTags.tileReset(tileIndex)),
-                                onResetClick = { onTileResetTapped(tileIndex) }
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun tileLeftOperandContentDescription(tile: TileUiState): String = if (tile.leftOperandLabel == "?") {
-    stringResource(R.string.tile_left_operand_hidden_content_description)
-} else {
-    stringResource(R.string.tile_left_operand_content_description, tile.leftOperandLabel)
-}
-
-@Composable
-private fun tileRightOperandContentDescription(tile: TileUiState): String = if (tile.rightOperandLabel == "?") {
-    stringResource(R.string.tile_right_operand_hidden_content_description)
-} else {
-    stringResource(R.string.tile_right_operand_content_description, tile.rightOperandLabel)
-}
-
-@Composable
-private fun tileOperatorContentDescription(tile: TileUiState): String = if (tile.operatorLabel == "?") {
-    stringResource(R.string.tile_operator_hidden_content_description)
-} else {
-    stringResource(
-        R.string.tile_operator_content_description,
-        tile.operatorAccessibilityLabel()
-    )
-}
-
-@Composable
-private fun TileUiState.operatorAccessibilityLabel(): String = when (operatorLabel) {
-    "+" -> stringResource(R.string.tile_operator_option_addition)
-    "×" -> stringResource(R.string.tile_operator_option_multiplication)
-    else -> operatorLabel
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun TileOperandSelectionSheet(
-    dialogUiState: TileOperandSelectionDialogUiState,
-    onDismiss: () -> Unit,
-    onConfirm: (Int) -> Unit
-) {
-    val operandSheetContentDescription = stringResource(R.string.tile_operand_dialog_title)
-
-    ModalBottomSheet(
-        modifier = Modifier
-            .testTag(GameScreenTestTags.TILE_OPERAND_SELECTOR)
-            .semantics {
-                contentDescription = operandSheetContentDescription
-            },
-        onDismissRequest = onDismiss
-    ) {
-        Text(
-            text = operandSheetContentDescription,
-            style = MaterialTheme.typography.titleMedium,
-            modifier = Modifier.padding(
-                start = TILE_OPERAND_SHEET_PADDING,
-                end = TILE_OPERAND_SHEET_PADDING,
-                bottom = 12.dp
-            )
-        )
-        LazyVerticalGrid(
-            columns = GridCells.Adaptive(minSize = TILE_OPERAND_SHEET_OPTION_MIN_WIDTH),
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(max = TILE_OPERAND_SHEET_MAX_HEIGHT)
-                .selectableGroup(),
-            contentPadding = PaddingValues(
-                start = TILE_OPERAND_SHEET_PADDING,
-                end = TILE_OPERAND_SHEET_PADDING,
-                bottom = TILE_OPERAND_SHEET_PADDING
-            ),
-            horizontalArrangement = Arrangement.spacedBy(TILE_OPERAND_SHEET_GRID_SPACING),
-            verticalArrangement = Arrangement.spacedBy(TILE_OPERAND_SHEET_GRID_SPACING)
-        ) {
-            items(
-                items = dialogUiState.availableOperands,
-                key = TileOperandOptionUiState::stripEntryId
-            ) { operand ->
-                val operandSelectionLabel = operand.value.toString()
-                val optionColors = operandOptionColors(
-                    enabled = operand.isSelectable,
-                    colorScheme = MaterialTheme.colorScheme
-                )
-
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = TILE_OPERAND_HINT_OVERLAY_LIFT),
-                    contentAlignment = Alignment.TopCenter
-                ) {
-                    Box {
-                        Surface(
-                            onClick = { onConfirm(operand.stripEntryId) },
-                            enabled = operand.isSelectable,
-                            modifier = Modifier
-                                .widthIn(
-                                    min = TILE_OPERAND_SHEET_OPTION_CARD_MIN_WIDTH,
-                                    max = TILE_OPERAND_SHEET_OPTION_CARD_MAX_WIDTH
-                                )
-                                .defaultMinSize(minHeight = TILE_OPERAND_SHEET_OPTION_MIN_HEIGHT)
-                                .testTag(GameScreenTestTags.tileOperandOption(operand.stripEntryId))
-                                .semantics {
-                                    contentDescription = operandSelectionLabel
-                                },
-                            shape = RoundedCornerShape(TILE_OPERAND_SHEET_OPTION_CORNER_RADIUS),
-                            color = optionColors.container,
-                            contentColor = optionColors.content,
-                            border = BorderStroke(
-                                width = 1.dp,
-                                color = optionColors.border
-                            )
-                        ) {
-                            Box(
-                                modifier = Modifier.fillMaxWidth(),
-                                contentAlignment = Alignment.Center
-                            ) {
-                                Text(
-                                    text = operandSelectionLabel,
-                                    modifier = Modifier.padding(horizontal = 14.dp, vertical = 9.dp),
-                                    style = MaterialTheme.typography.titleMedium,
-                                    fontWeight = FontWeight.SemiBold
-                                )
-                            }
-                        }
-                        OperandUsageHintBadge(
-                            operator = Operator.ADDITION,
-                            usageState = operand.usageStateFor(Operator.ADDITION),
-                            stripEntryId = operand.stripEntryId,
-                            enabled = operand.isSelectable,
-                            modifier = Modifier
-                                .zIndex(1f)
-                                .align(Alignment.TopStart)
-                                .padding(start = TILE_OPERAND_HINT_EDGE_INSET)
-                                .offset(y = -TILE_OPERAND_HINT_OVERLAY_LIFT)
-                        )
-                        OperandUsageHintBadge(
-                            operator = Operator.MULTIPLICATION,
-                            usageState = operand.usageStateFor(Operator.MULTIPLICATION),
-                            stripEntryId = operand.stripEntryId,
-                            enabled = operand.isSelectable,
-                            modifier = Modifier
-                                .zIndex(1f)
-                                .align(Alignment.TopEnd)
-                                .padding(end = TILE_OPERAND_HINT_EDGE_INSET)
-                                .offset(y = -TILE_OPERAND_HINT_OVERLAY_LIFT)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun OperandUsageHintBadge(
-    modifier: Modifier = Modifier,
-    operator: Operator,
-    usageState: OperandUsageHintState,
-    stripEntryId: Int,
-    enabled: Boolean = true
-) {
-    val colorScheme = MaterialTheme.colorScheme
-    val hintContentDescription = when (operator) {
-        Operator.Addition -> stringResource(R.string.tile_operand_usage_addition_hint)
-        Operator.Multiplication -> stringResource(R.string.tile_operand_usage_multiplication_hint)
-        Operator.Hidden -> error("Hidden operator does not expose operand usage hints.")
-    }
-    val hintStateDescription = stringResource(usageState.stateDescriptionResId)
-    val resolvedColors = baseHintBadgeColors(
-        usageState = usageState,
-        operator = operator,
-        colorScheme = colorScheme
-    ).let { colors ->
-        if (enabled) colors else colors.disabled(colorScheme)
-    }
-
-    Surface(
-        modifier = modifier
-            .testTag(GameScreenTestTags.tileOperandUsageHint(stripEntryId, operator))
-            .semantics {
-                contentDescription = hintContentDescription
-                stateDescription = hintStateDescription
-            },
-        shape = RoundedCornerShape(TILE_OPERAND_HINT_CORNER_RADIUS),
-        color = resolvedColors.container,
-        contentColor = resolvedColors.content,
-        border = BorderStroke(width = 1.dp, color = resolvedColors.border)
-    ) {
-        Text(
-            text = operator.symbol,
-            modifier = Modifier.padding(
-                horizontal = TILE_OPERAND_HINT_HORIZONTAL_PADDING,
-                vertical = TILE_OPERAND_HINT_VERTICAL_PADDING
-            ),
-            style = MaterialTheme.typography.labelSmall,
-            fontWeight = TILE_OPERAND_HINT_TEXT_WEIGHT
-        )
-    }
-}
-
-private data class HintBadgeColors(val container: Color, val content: Color, val border: Color)
-
-private fun baseHintBadgeColors(
-    usageState: OperandUsageHintState,
-    operator: Operator,
-    colorScheme: ColorScheme
-): HintBadgeColors = when (usageState) {
-    OperandUsageHintState.AVAILABLE -> HintBadgeColors(
-        container = colorScheme.surfaceContainerHighest,
-        content = colorScheme.onSurfaceVariant,
-        border = colorScheme.outline
-    )
-
-    OperandUsageHintState.USED -> when (operator) {
-        Operator.Addition,
-        Operator.Multiplication -> HintBadgeColors(
-            container = colorScheme.primaryContainer,
-            content = colorScheme.onPrimaryContainer,
-            border = colorScheme.primary
-        )
-
-        Operator.Hidden -> error("Hidden operator does not expose operand usage hints.")
-    }
-}
-
-private fun HintBadgeColors.disabled(colorScheme: ColorScheme): HintBadgeColors = HintBadgeColors(
-    container = lerp(container, colorScheme.surfaceContainerHigh, 0.35f),
-    content = lerp(content, colorScheme.onSurface, 0.20f),
-    border = lerp(border, colorScheme.outline, 0.35f)
-)
-
-private data class OperandOptionColors(val container: Color, val content: Color, val border: Color)
-
-private fun operandOptionColors(enabled: Boolean, colorScheme: ColorScheme): OperandOptionColors = if (enabled) {
-    OperandOptionColors(
-        container = colorScheme.surface,
-        content = colorScheme.onSurface,
-        border = colorScheme.outline
-    )
-} else {
-    OperandOptionColors(
-        container = colorScheme.surfaceContainerHigh,
-        content = colorScheme.onSurfaceVariant,
-        border = colorScheme.outline
-    )
-}
-
-private fun TileOperandOptionUiState.usageStateFor(operator: Operator): OperandUsageHintState = when (operator) {
-    Operator.Addition -> when {
-        additionUsed -> OperandUsageHintState.USED
-        else -> OperandUsageHintState.AVAILABLE
-    }
-
-    Operator.Multiplication -> when {
-        multiplicationUsed -> OperandUsageHintState.USED
-        else -> OperandUsageHintState.AVAILABLE
-    }
-
-    Operator.Hidden -> error("Hidden operator does not expose operand usage hints.")
-}
-
-private enum class OperandUsageHintState(val stateDescriptionResId: Int) {
-    AVAILABLE(R.string.tile_operand_usage_state_available),
-    USED(R.string.tile_operand_usage_state_used)
-}
-
-@Composable
-private fun StripSection(
-    stripItems: List<StripItemUiState>,
-    onStripItemTapped: (Int) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val stripContentDescription = stringResource(R.string.strip_content_description)
-
-    Surface(
-        modifier = modifier
-            .testTag(GameScreenTestTags.STRIP)
-            .semantics {
-                contentDescription = stripContentDescription
-            },
-        color = MaterialTheme.colorScheme.surfaceVariant,
-        shape = RoundedCornerShape(24.dp)
-    ) {
-        BoxWithConstraints(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    horizontal = STRIP_HORIZONTAL_PADDING,
-                    vertical = STRIP_VERTICAL_PADDING
-                )
-        ) {
-            val chipCount = stripItems.size
-            val chipWidth = calculateStripChipWidth(
-                availableWidth = maxWidth,
-                chipCount = chipCount
-            )
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(STRIP_CHIP_SPACING)
-            ) {
-                stripItems.forEachIndexed { index, stripItem ->
-                    val stripItemContentDescription = when (stripItem.visualStyle) {
-                        StripItemVisualStyle.KNOWN -> stringResource(
-                            R.string.strip_item_known_content_description,
-                            stripItem.label
-                        )
-
-                        StripItemVisualStyle.HIDDEN -> stringResource(
-                            R.string.strip_item_hidden_content_description
-                        )
-
-                        StripItemVisualStyle.PLAYER_ENTERED -> stringResource(
-                            R.string.strip_item_player_entered_content_description,
-                            stripItem.label
-                        )
-                    }
-
-                    AvailableNumberChip(
-                        label = stripItem.label,
-                        modifier = Modifier
-                            .width(chipWidth)
-                            .testTag(GameScreenTestTags.stripItem(index)),
-                        contentDescription = stripItemContentDescription,
-                        style = when (stripItem.visualStyle) {
-                            StripItemVisualStyle.KNOWN -> AvailableNumberChipStyle.KNOWN
-                            StripItemVisualStyle.HIDDEN -> AvailableNumberChipStyle.HIDDEN
-                            StripItemVisualStyle.PLAYER_ENTERED -> AvailableNumberChipStyle.PLAYER_ENTERED
-                        },
-                        onClick = if (stripItem.isEntryEnabled) {
-                            { onStripItemTapped(index) }
-                        } else {
-                            null
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun StripItemEntryDialog(
-    dialogUiState: StripItemEntryDialogUiState,
-    onDismiss: () -> Unit,
-    onConfirm: (Int) -> Unit
-) {
-    var enteredValue by rememberSaveable(
-        dialogUiState.stripItemIndex,
-        dialogUiState.mode,
-        dialogUiState.initialValue
-    ) {
-        mutableStateOf(dialogUiState.initialValue)
-    }
-    val parsedValue = enteredValue.toIntOrNull()
-    val confirmedValue = parsedValue?.takeIf { it in dialogUiState.validRange }
-    val isInputInvalid = enteredValue.isNotEmpty() && confirmedValue == null
-    val validRangeText = dialogUiState.validRange.maximumValue?.let { maximumValue ->
-        stringResource(
-            R.string.strip_entry_valid_range_bounded,
-            dialogUiState.validRange.minimumValue,
-            maximumValue
-        )
-    } ?: stringResource(
-        R.string.strip_entry_valid_range_unbounded,
-        dialogUiState.validRange.minimumValue
-    )
-
-    AlertDialog(
-        modifier = Modifier.testTag(GameScreenTestTags.STRIP_ENTRY_DIALOG),
-        onDismissRequest = onDismiss,
-        title = {
-            Text(text = stringResource(R.string.strip_entry_dialog_title))
-        },
-        text = {
-            OutlinedTextField(
-                value = enteredValue,
-                onValueChange = { newValue ->
-                    enteredValue = newValue.filter(Char::isDigit)
-                },
-                modifier = Modifier.testTag(GameScreenTestTags.STRIP_ENTRY_INPUT),
-                label = {
-                    Text(text = stringResource(R.string.strip_entry_input_label))
-                },
-                supportingText = {
-                    Text(
-                        text = validRangeText,
-                        modifier = Modifier.testTag(GameScreenTestTags.STRIP_ENTRY_RANGE)
-                    )
-                },
-                isError = isInputInvalid,
-                singleLine = true,
-                keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number)
-            )
-        },
-        confirmButton = {
-            Button(
-                onClick = { onConfirm(confirmedValue ?: return@Button) },
-                enabled = confirmedValue != null,
-                modifier = Modifier.testTag(GameScreenTestTags.STRIP_ENTRY_CONFIRM)
-            ) {
-                Text(text = stringResource(R.string.confirm))
-            }
-        },
-        dismissButton = {
-            TextButton(
-                onClick = onDismiss,
-                modifier = Modifier.testTag(GameScreenTestTags.STRIP_ENTRY_CANCEL)
-            ) {
-                Text(text = stringResource(R.string.cancel))
-            }
-        }
-    )
-}
-
-@Composable
-private fun TileOperatorSelectionMenu(
-    dialogUiState: TileOperatorSelectionDialogUiState,
-    onDismiss: () -> Unit,
-    onConfirm: (Operator) -> Unit
-) {
-    val operatorMenuContentDescription = stringResource(R.string.tile_operator_dialog_title)
-
-    DropdownMenu(
-        expanded = true,
-        onDismissRequest = onDismiss,
-        modifier = Modifier
-            .testTag(GameScreenTestTags.TILE_OPERATOR_SELECTOR)
-            .semantics {
-                contentDescription = operatorMenuContentDescription
-            },
-        offset = DpOffset(x = 0.dp, y = 4.dp)
-    ) {
-        Row(
-            modifier = Modifier
-                .padding(TILE_OPERATOR_MENU_PADDING)
-                .selectableGroup(),
-            horizontalArrangement = Arrangement.spacedBy(TILE_OPERATOR_MENU_OPTION_SPACING)
-        ) {
-            dialogUiState.availableOperators.forEach { operator ->
-                val isSelected = dialogUiState.initialOperator == operator
-                val operatorSelectionLabel = operator.selectionLabel()
-
-                Surface(
-                    onClick = { onConfirm(operator) },
-                    modifier = Modifier
-                        .defaultMinSize(minWidth = 0.dp)
-                        .testTag(GameScreenTestTags.tileOperatorOption(operator))
-                        .semantics {
-                            contentDescription = operatorSelectionLabel
-                            selected = isSelected
-                        },
-                    shape = RoundedCornerShape(TILE_OPERATOR_MENU_CORNER_RADIUS),
-                    color = if (isSelected) {
-                        MaterialTheme.colorScheme.primaryContainer
-                    } else {
-                        MaterialTheme.colorScheme.surface
-                    },
-                    contentColor = if (isSelected) {
-                        MaterialTheme.colorScheme.onPrimaryContainer
-                    } else {
-                        MaterialTheme.colorScheme.onSurface
-                    },
-                    border = BorderStroke(
-                        width = 1.dp,
-                        color = if (isSelected) {
-                            MaterialTheme.colorScheme.primary
-                        } else {
-                            MaterialTheme.colorScheme.outlineVariant
-                        }
-                    )
-                ) {
-                    Text(
-                        text = operator.symbol,
-                        modifier = Modifier.padding(
-                            horizontal = TILE_OPERATOR_MENU_OPTION_HORIZONTAL_PADDING,
-                            vertical = TILE_OPERATOR_MENU_OPTION_VERTICAL_PADDING
-                        ),
-                        fontWeight = if (isSelected) {
-                            FontWeight.SemiBold
-                        } else {
-                            FontWeight.Normal
-                        }
-                    )
-                }
-            }
-        }
-    }
-}
-
-private fun calculateStripChipWidth(availableWidth: Dp, chipCount: Int): Dp {
-    val totalSpacing = STRIP_CHIP_SPACING * (chipCount - 1)
-    return (availableWidth - totalSpacing) / chipCount
-}
-
-private fun calculateBoardColumnCount(availableWidth: Dp): Int {
-    val columnsThatFit = (
-        (availableWidth.value + BOARD_TILE_SPACING.value) /
-            (BOARD_TILE_MIN_WIDTH.value + BOARD_TILE_SPACING.value)
-        ).toInt()
-
-    return columnsThatFit.coerceIn(1, BOARD_MAX_VISUAL_COLUMN_COUNT)
-}
-
-private fun calculateBoardTileWidth(availableWidth: Dp, visualColumnCount: Int): Dp {
-    val totalSpacing = BOARD_TILE_SPACING * (visualColumnCount - 1)
-    val availableTileWidth = (availableWidth - totalSpacing) / visualColumnCount
-
-    return availableTileWidth.coerceIn(BOARD_TILE_MIN_WIDTH, BOARD_TILE_MAX_WIDTH)
-}
-
-@Composable
-private fun Operator.selectionLabel(): String = when (this) {
-    Operator.Addition -> stringResource(R.string.tile_operator_option_addition)
-    Operator.Multiplication -> stringResource(R.string.tile_operator_option_multiplication)
-    Operator.Hidden -> error("Hidden operator is not a selectable option.")
-}
-
-@Composable
-private fun PuzzleOutcomeUiState.Invalid.message(): String = when (completionState) {
-    PuzzleCompletionState.INCORRECT_TILES -> stringResource(R.string.puzzle_outcome_invalid_tiles_message)
-    PuzzleCompletionState.MISSING_STRIP_ENTRY_IDENTITIES ->
-        stringResource(R.string.puzzle_outcome_missing_identities_message)
-    PuzzleCompletionState.MISMATCHED_SUM_PRODUCT_PAIRINGS ->
-        stringResource(R.string.puzzle_outcome_mismatched_pairings_message)
-    PuzzleCompletionState.INVALID_STRIP_ENTRY_USAGE ->
-        stringResource(R.string.puzzle_outcome_invalid_usage_message)
-    PuzzleCompletionState.INCOMPLETE,
-    PuzzleCompletionState.SOLVED -> error("Invalid outcome must represent a completed unsolved puzzle.")
 }
 
 @Preview(showBackground = true)
