+++ app/src/main/java/org/cescfe/numpairs/ui/screen/GameScreenSections.kt
@@ -1,0 +1,166 @@
+package org.cescfe.numpairs.ui.screen
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.contentDescription
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.unit.dp
+import org.cescfe.numpairs.R
+import org.cescfe.numpairs.domain.puzzle.Operator
+import org.cescfe.numpairs.ui.components.AvailableNumberChip
+import org.cescfe.numpairs.ui.components.AvailableNumberChipStyle
+import org.cescfe.numpairs.ui.components.PuzzleTile
+
+@Composable
+internal fun BoardSection(
+    tiles: List<TileUiState>,
+    onTileLeftOperandTapped: (Int) -> Unit,
+    onTileRightOperandTapped: (Int) -> Unit,
+    onTileOperatorTapped: (Int) -> Unit,
+    onTileResetTapped: (Int) -> Unit,
+    tileOperatorSelectionDialog: TileOperatorSelectionDialogUiState?,
+    onTileOperatorSelectionDismissed: () -> Unit,
+    onTileOperatorSelectionConfirmed: (Operator) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val boardContentDescription = stringResource(R.string.board_content_description)
+
+    BoxWithConstraints(
+        modifier = modifier
+            .testTag(GameScreenTestTags.BOARD)
+            .semantics {
+                contentDescription = boardContentDescription
+            }
+    ) {
+        val visualColumnCount = calculateBoardColumnCount(maxWidth)
+        val tileWidth = calculateBoardTileWidth(
+            availableWidth = maxWidth,
+            visualColumnCount = visualColumnCount
+        )
+        val visualRows = tiles.withIndex().toList().chunked(visualColumnCount)
+
+        Column(
+            verticalArrangement = Arrangement.spacedBy(BOARD_TILE_SPACING)
+        ) {
+            visualRows.forEach { row ->
+                Box(
+                    modifier = Modifier.fillMaxWidth(),
+                    contentAlignment = Alignment.Center
+                ) {
+                    Row(
+                        horizontalArrangement = Arrangement.spacedBy(BOARD_TILE_SPACING)
+                    ) {
+                        row.forEach { indexedTile ->
+                            val tileIndex = indexedTile.index
+                            val tile = indexedTile.value
+                            val tileOperatorSelectionUiState = tileOperatorSelectionDialog
+                                ?.takeIf { dialogUiState -> dialogUiState.tileIndex == tileIndex }
+
+                            PuzzleTile(
+                                tile = tile,
+                                modifier = Modifier
+                                    .testTag(GameScreenTestTags.tile(tileIndex))
+                                    .width(tileWidth)
+                                    .wrapContentHeight(),
+                                leftOperandModifier = Modifier.testTag(GameScreenTestTags.tileLeftOperand(tileIndex)),
+                                leftOperandContentDescription = tileLeftOperandContentDescription(tile),
+                                onLeftOperandClick = { onTileLeftOperandTapped(tileIndex) },
+                                operatorModifier = Modifier.testTag(GameScreenTestTags.tileOperator(tileIndex)),
+                                operatorContentDescription = tileOperatorContentDescription(tile),
+                                onOperatorClick = { onTileOperatorTapped(tileIndex) },
+                                operatorOverlay = {
+                                    tileOperatorSelectionUiState?.let { dialogUiState ->
+                                        TileOperatorSelectionMenu(
+                                            dialogUiState = dialogUiState,
+                                            onDismiss = onTileOperatorSelectionDismissed,
+                                            onConfirm = onTileOperatorSelectionConfirmed
+                                        )
+                                    }
+                                },
+                                rightOperandModifier = Modifier.testTag(GameScreenTestTags.tileRightOperand(tileIndex)),
+                                rightOperandContentDescription = tileRightOperandContentDescription(tile),
+                                onRightOperandClick = { onTileRightOperandTapped(tileIndex) },
+                                resetModifier = Modifier.testTag(GameScreenTestTags.tileReset(tileIndex)),
+                                onResetClick = { onTileResetTapped(tileIndex) }
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Composable
+internal fun StripSection(
+    stripItems: List<StripItemUiState>,
+    onStripItemTapped: (Int) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val stripContentDescription = stringResource(R.string.strip_content_description)
+
+    Surface(
+        modifier = modifier
+            .testTag(GameScreenTestTags.STRIP)
+            .semantics {
+                contentDescription = stripContentDescription
+            },
+        color = MaterialTheme.colorScheme.surfaceVariant,
+        shape = RoundedCornerShape(24.dp)
+    ) {
+        BoxWithConstraints(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(
+                    horizontal = STRIP_HORIZONTAL_PADDING,
+                    vertical = STRIP_VERTICAL_PADDING
+                )
+        ) {
+            val chipCount = stripItems.size
+            val chipWidth = calculateStripChipWidth(
+                availableWidth = maxWidth,
+                chipCount = chipCount
+            )
+
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(STRIP_CHIP_SPACING)
+            ) {
+                stripItems.forEachIndexed { index, stripItem ->
+                    AvailableNumberChip(
+                        label = stripItem.label,
+                        modifier = Modifier
+                            .width(chipWidth)
+                            .testTag(GameScreenTestTags.stripItem(index)),
+                        contentDescription = stripItemContentDescription(stripItem),
+                        style = when (stripItem.visualStyle) {
+                            StripItemVisualStyle.KNOWN -> AvailableNumberChipStyle.KNOWN
+                            StripItemVisualStyle.HIDDEN -> AvailableNumberChipStyle.HIDDEN
+                            StripItemVisualStyle.PLAYER_ENTERED -> AvailableNumberChipStyle.PLAYER_ENTERED
+                        },
+                        onClick = if (stripItem.isEntryEnabled) {
+                            { onStripItemTapped(index) }
+                        } else {
+                            null
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
